+++ src/test/java/org/elite/ir/IRInterpreterTest.java
@@ -83,7 +83,7 @@ class IRInterpreterTest {
 
     // ── helpers ──
 
-    private static IRFunction buildFn(String name, long[] code, int paramCount) {
+    private static IRFunction buildFn(String name, int[] code, int paramCount) {
         IRFunction irf = new IRFunction(name, paramCount);
         irf.populate(code, 2, new int[]{0}, new Object[]{0,1}, null, null);
         return irf;
